% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_values}
\alias{sim_values}
\title{Simulate Values}
\usage{
sim_values(n, dist = "Normal", ...)
}
\arguments{
\item{n}{numeric value for number of values to be returned}

\item{dist}{string value of distribution to sample from}

\item{...}{additional arguments for defining distributions. These
include: normal (mean, sd), poisson (lambda) and binomial (size, prob)}
}
\value{
Vector of simulated values
}
\description{
A function used for
}
\examples{
sim_values(10, "Normal", mean = 5)

}

